refactor: modularize UI components into ui/ directory
Break down the monolithic ui_components.py (1,009 lines) into a maintainable
modular structure under src/squiggy/ui/ directory.

Changes:
- Extract FileInfoPanel to ui/file_info.py (95 lines)
- Extract PlotOptionsPanel to ui/plot_options.py (174 lines)
- Extract AdvancedOptionsPanel to ui/advanced_options.py (163 lines)
- Extract SearchPanel to ui/search.py (119 lines)
- Extract ModificationsPanel to ui/modifications.py (417 lines)
- Create ui/__init__.py with backward-compatible re-exports
- Update imports in viewer.py: from .ui_components → from .ui
- Update imports in tests/test_modifications.py
- Remove old ui_components.py file

Benefits:
- Smaller, focused files (100-200 lines each, except ModificationsPanel at 417)
- Easier navigation and maintenance
- Clearer separation of concerns
- Reduced merge conflicts
- Easier testing of individual components

All 414 tests passing, no functional changes.
